infer start time before writing to db
