+++ src/replays/database_replay_provider/record_mapper.ts
@@ -1,6 +1,5 @@
 import type { FileResult, PlayerInfo } from "@replays/types";
 import type { Game, Player, Replay } from "database/schema";
-import moment from "moment";
 import path from "path";
 
 function mapPlayerRecordToPlayerInfo(player: Player): PlayerInfo {
@@ -29,7 +28,7 @@ export function mapGameRecordToFileResult(gameRecord: Game & Replay, playerRecor
       players: playerRecords.map(mapPlayerRecordToPlayerInfo),
       isTeams: Boolean(gameRecord.is_teams),
       stageId: gameRecord.stage,
-      startTime: inferStartTime(gameRecord.start_time, gameRecord.file_name, gameRecord.birth_time),
+      startTime: gameRecord.start_time,
       platform: gameRecord.platform,
       consoleNickname: gameRecord.console_nickname,
       mode: gameRecord.mode,
@@ -38,44 +37,4 @@ export function mapGameRecordToFileResult(gameRecord: Game & Replay, playerRecor
       startingTimerSeconds: gameRecord.starting_timer_secs,
     },
   };
-}
-
-function inferStartTime(gameStartAt: string | null, fileName: string, fileBirthTime: string | null): string | null {
-  let startAt = convertToDateAndTime(gameStartAt);
-  if (startAt) {
-    return startAt.toISOString();
-  }
-
-  startAt = inferDateFromFilename(fileName);
-  if (startAt) {
-    return startAt.toISOString();
-  }
-
-  startAt = convertToDateAndTime(fileBirthTime);
-  return startAt?.toISOString() ?? null;
-}
-
-function convertToDateAndTime(dateTimeString: moment.MomentInput): moment.Moment | null {
-  if (dateTimeString == null) {
-    return null;
-  }
-
-  const asMoment = moment(dateTimeString);
-  if (asMoment.isValid()) {
-    return asMoment.local();
-  }
-
-  return null;
-}
-
-function inferDateFromFilename(fileName: string): moment.Moment | null {
-  const timeReg = /\d{8}T\d{6}/g;
-  const filenameTime = fileName.match(timeReg);
-
-  if (filenameTime == null) {
-    return null;
-  }
-
-  const time = moment(filenameTime[0]).local();
-  return time;
 }
